+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+* Flatten udnssdk.Response to mere http.Response
 
 ## [1.1.1] - 2016-05-27
 ### Fixed